Delta time implementation

+++ include/ac_bee.h
@@ -30,9 +30,9 @@ struct bee_actor_s {
     s16 add_angle;
     s16 fly_angle[2];
     s16 base_angle;
-    int catch_delay_frames;
+    f32 catch_delay_frames;
     ACTOR* insect_actor;
-    int disappear_timer;
+    f32 disappear_timer;
 };
 
 extern ACTOR_PROFILE Bee_Profile;
