+++ include/ac_npc_rtc.h
@@ -22,9 +22,9 @@ struct actor_npc_rtc_s {
     int unk9A4;
     int think_idx;
     int counter;
-    int fade_timer;
-    int bgm_timer;
-    int first_talk_timer;
+    f32 fade_timer;
+    f32 bgm_timer;
+    f32 first_talk_timer;
 };
 
 extern ACTOR_PROFILE Npc_Rtc_Profile;
