+++ include/m_player.h
@@ -1625,7 +1625,7 @@ typedef struct player_main_swing_axe_s {
     /* 0x14 */ int tree_ut_z;
     /* 0x18 */ int bee_flag;
     /* 0x1C */ s16 bee_angle_y;
-    /* 0x20 */ int bee_counter;
+    /* 0x20 */ f32 bee_counter;
 } mPlayer_main_swing_axe_c;
 
 typedef struct player_main_reflect_axe_s {
@@ -1843,7 +1843,7 @@ typedef struct player_main_shake_tree_s {
     int tree_ut_z;
     int bee_flag;
     s16 bee_angle_y;
-    int bee_spawn_timer;
+    f32 bee_spawn_timer;
 } mPlayer_main_shake_tree_c;
 
 typedef struct player_main_struggle_pitfall_s {
